LayoutTests/imported/w3c/IndexedDB-private-browsing/idbindex_openCursor-expected.txt


PASS IDBIndex.openCursor() - throw InvalidStateError when the index is deleted
